fixed various clang tidy warnings

+++ .src/unsafe_utilities.cpp
@@ -194,7 +194,7 @@ namespace jluna::unsafe
                 types.at(i) = (unsafe::Value*) jl_int64_type;
 
             auto* tuple_type = jl_apply_tuple_type_v(types.data(), types.size());
-            auto* tuple = jl_new_struct(tuple_type, jl_box_int64(one_d));
+            auto* tuple = jl_new_struct(tuple_type, jl_box_int64(static_cast<Int64>(one_d)));
             auto* res = jl_reshape_array(jl_apply_array_type(unsafe::call(array_value_t, array), 1), array, tuple);
             override_array(array, res);
             gc_unpause;
@@ -219,7 +219,7 @@ namespace jluna::unsafe
                 types.at(i) = (unsafe::Value*) jl_int64_type;
 
             auto* tuple_type = jl_apply_tuple_type_v(types.data(), types.size());
-            auto* tuple = jl_new_struct(tuple_type, jl_box_int64(one_d), jl_box_int64(two_d));
+            auto* tuple = jl_new_struct(tuple_type, jl_box_int64(static_cast<Int64>(one_d)), jl_box_int64(static_cast<Int64>(two_d)));
             auto* res = jl_reshape_array(jl_apply_array_type(unsafe::call(array_value_t, array), 2), array, tuple);
             override_array(array, res);
             gc_unpause;
